refactored loans and credit cards screens
