+++ mobile-app/app/src/main/java/com/semirsuljevic/raiffaisenmobileapp/ui/navigation/navigator.kt
@@ -23,6 +23,8 @@ import com.semirsuljevic.raiffaisenmobileapp.ui.screens.intro_info.*
 import com.semirsuljevic.raiffaisenmobileapp.ui.screens.intro_navbar.IntroNavbar
 import com.semirsuljevic.raiffaisenmobileapp.ui.screens.intro_navbar.LoginScreen
 import com.semirsuljevic.raiffaisenmobileapp.ui.screens.intro_navbar.intro_more.*
+import com.semirsuljevic.raiffaisenmobileapp.ui.screens.intro_navbar.intro_products.CreditCardsScreen
+import com.semirsuljevic.raiffaisenmobileapp.ui.screens.intro_navbar.intro_products.LoansScreen
 import com.semirsuljevic.raiffaisenmobileapp.ui.screens.intro_navbar.locations_filter.LocationsFilterScreen
 import com.semirsuljevic.raiffaisenmobileapp.ui.theme.Gray200
 import com.semirsuljevic.raiffaisenmobileapp.ui.view_models.SecureSharedPref
